requirements.txt
playwright==1.40.0
pandas==2.1.4
python-dotenv==1.0.0
schedule==1.2.0
google-cloud-storage==2.14.0
google-cloud-bigquery==3.14.1
